# src/Ai.Tlbx.MidTerm/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MidTerm API",
    "description": "MidTerm terminal multiplexer API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/files/register": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/check": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/view": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/download": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/resolve": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResolveResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectoryEntry": {
        "required": [
          "name",
          "isDirectory",
          "size",
          "modified",
          "mimeType"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DirectoryListResponse": {
        "required": [
          "path",
          "entries"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectoryEntry"
            }
          }
        }
      },
...
      "FileCheckRequest": {
        "required": [
          "paths"
        ],
        "type": "object",
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileCheckResponse": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FilePathInfo"
            }
          }
        }
      },
      "FilePathInfo": {
        "required": [
          "exists",
          "size",
          "isDirectory",
          "mimeType",
          "modified",
          "isText"
        ],
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isText": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "FileRegisterRequest": {
        "required": [
          "sessionId",
          "paths"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileResolveResponse": {
        "required": [
          "exists",
          "resolvedPath",
          "isDirectory",
          "size",
          "mimeType",
          "modified",
          "isText"
        ],
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "resolvedPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDirectory": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isText": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    }
  },
...
}